# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cohort Discovery Service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/queries": {
            "get": {
                "tags": [
                    "Queries"
                ],
                "summary": "List queries for the authenticated user",
                "operationId": "83d4c8cfd0361ba3173db16ba127847d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of queries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Query"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Queries"
                ],
                "summary": "Create and submit a new query",
                "operationId": "e8f157d94686adf23d54ba8612feda6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Query"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created and submitted query result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Query"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/queries/{key}": {
            "get": {
                "tags": [
                    "Queries"
                ],
                "summary": "Get a single query by id or pid",
                "operationId": "c45855fb50b838b4af26f643ce0cc2ff",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Database id or public pid of the query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Query"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Queries"
                ],
                "summary": "Update an existing query (by id or pid)",
                "operationId": "b50fe1ac92d51202351a06cdce919156",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Database id or public pid of the query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "col_abc123"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Query"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Query"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Queries"
                ],
                "summary": "Delete a query by id or pid",
                "operationId": "36839b2d4517ddd975f93d99ce33e1e0",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Database id or public pid of the query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/queries/translate/{context}": {
            "post": {
                "tags": [
                    "Queries"
                ],
                "summary": "Translate a raw query definition into a specific query context format, without submitting it",
                "operationId": "56e8758501e6a3db9d52d04819bcc242",
                "parameters": [
                    {
                        "name": "context",
                        "in": "path",
                        "description": "Query context slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "bunny",
                                "beacon"
                            ],
                            "example": "bunny"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "definition"
                                ],
                                "properties": {
                                    "definition": {
                                        "description": "Structured query definition (same shape as Query::definition)",
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translated query context"
                    },
                    "422": {
                        "description": "Validation error, or unsupported context"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/queries/{pid}/download": {
            "get": {
                "tags": [
                    "Queries"
                ],
                "summary": "Download query results for a saved query",
                "operationId": "28cc28c8a59bb1ddf465e2f93d3ce0bd",
                "parameters": [
                    {
                        "name": "pid",
                        "in": "path",
                        "description": "Public pid of the saved query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "qry_abc123"
                        }
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "description": "Output format (csv|json)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "csv"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streamed file download (binary)",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Query": {
                "title": "Query",
                "required": [
                    "pid",
                    "name",
                    "definition"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pid": {
                        "type": "string",
                        "example": "qry_abc123"
                    },
                    "name": {
                        "type": "string",
                        "example": "Cardiology cohort query"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "definition": {
                        "description": "Structured query definition (array of rule objects)",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-06T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}